@@ -5,18 +5,18 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.5.2</version>
+		<version>3.0.5</version>
 		<relativePath/>
 	</parent>
 	<groupId>dev.lydtech</groupId>
 	<artifactId>ctf-example-service</artifactId>
-	<version>1.6.0</version>
+	<version>2.0.0</version>
 	<name>ctf-example-service</name>
 	<description>Example service that showcases the component-test-framework for component testing</description>
 	<properties>
-		<java.version>11</java.version>
-		<spring.boot.version>2.5.2</spring.boot.version>
-		<spring.kafka.version>2.7.8</spring.kafka.version>
+		<java.version>17</java.version>
+		<maven.compiler.source>17</maven.compiler.source>
+		<maven.compiler.target>17</maven.compiler.target>
 
 		<!-- Allows control of whether to leave TestContainers up after component test run -->
 		<containers.stayup>false</containers.stayup>
